Naive Bayes e-mail identifier
We have a set of emails, half of which were written by one person and
the other half by another person at the same company . Our objective is
to classify the emails as written by one person or the other based only
on the text of the email. We are gonna use Naive Bayes. We also are
interested on checking which part is faster, the training or the
predicting section…we’ll see!